♻️ separate Agent thread route and DTO types
