+++ libs/frontend/state/conversation/agent-threads/src/lib/opencrane-agent-thread.mapper.types.ts
@@ -1,0 +1,4 @@
+import type { paths } from "@opencrane/contracts";
+
+/** Generated success DTO for one exact authorized Agent-thread route. */
+export type AgentThreadSnapshotDto = paths["/me/conversations/{parentConversationId}/agent-threads/{childConversationId}"]["get"]["responses"][200]["content"]["application/json"]["agentThread"];
