+++ libs/frontend/state/conversation/agent-threads/src/lib/opencrane-agent-thread.validator.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import type { AgentThreadSnapshotDto } from "./opencrane-agent-thread.mapper.js";
+import type { AgentThreadSnapshotDto } from "./opencrane-agent-thread.mapper.types.js";
 
 /** Bounded non-empty server text that is safe to retain as display input. */
 const _Text = z.string().min(1).max(20_000);
